Absorb `SimpleSlotWorker` into `SubspaceSlotWorker`
